Convert hyphen to en dash in badge title text
An en dash ‘–’ (or an em dash ‘—’) is more typographically correct than a hyphen ‘-’ when used to set off another phrase. It will make the text a tiny bit easier to read.

I also converted the hyphens in some other text: the explanations of the badges in gh_pages/index.md, and the beginning of gh_pages/history.md.

This change does not include generating new folders in badges/ and gh_pages/badges/ by running the fabfile. It does, however, update the following files with the markup that would be generated, so you don’t have to remember to update them yourself:

- the badge at the top of README.md
- the What It Looks Like section of gh_pages/index.md
- gh_pages/fabfile.py