Added a formal ontology based on SKOS.
In order to allow repostatus.org to be used in the linked open data
world, it would help to have the concepts expressed in a way that allows
linking with the wider semantic web. This it my attempt at doing so in
an as inobtrusive way as possible. I opted for JSON-LD as the
serialisation format, which is easily parsable and even retains some
value even if you were ignore the context that maps it to an RDF model.

I mapped the all of the concepts defined in repostatus to SKOS Concepts
(https://www.w3.org/TR/skos-reference/), as that is a common and
relatively simple way to organize knowledge. The concepts together form
a ConceptScheme. I added opengraph vocabulary to link to the badge image
(as SKOS has no vocabulary for that).

Ideally this ontology.jsonld file should be served on repostatus.org if
the client sends an HTTP request with header `Accept: application/json+ld` (content
negotation).